pyproject.toml
[tool.poetry]
build = 'build.py'
name = "buvar"
version = "0.9.1"
description = "General purpose asyncio service loader and dependency injector"
readme = "README.rst"
authors = ["Oliver Berger <diefans@gmail.com>"]
license = "Apache-2.0"
include = ['COPYRIGHT', 'tests/*', 'examples/*']

[tool.poetry.scripts]
buvar = 'buvar.cli:main'

[tool.poetry.dependencies]
python = "^3.6"
orjson = "^2.0"
tomlkit = "^0.5.3"
toml = "^0.10"
multidict = "^4.5"
structlog = "^19.1"
attrs = "^19.1"
typing_inspect = "^0.4.0"
cattrs = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pytest-randomly = "^3.0"
pytest-watch = "^4.2"
pytest-cache = "^1.0"
pytest-mock = "^1.10"
mock = "^3.0"
pylava = "^0.2.2"
pylava-pylint = "^0.0.3"
pycodestyle = "^2.5"
pydocstyle = "^3.0"
ipython = "^7.5"
isort = "^4.3"
pdbpp = "^0.10.0"
tox = "^3.12"
pytest-cov = "^2.7"
pytest-asyncio = "^0.10.0"
aiohttp = "^3.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
